Add parallel p4est tests
