+++ test/test_mpi_p4est.jl
@@ -1,0 +1,40 @@
+module TestExamplesMPIP4estMesh
+
+using Test
+using Trixi
+
+include("test_trixi.jl")
+
+# pathof(Trixi) returns /path/to/Trixi/src/Trixi.jl, dirname gives the parent directory
+const EXAMPLES_DIR = joinpath(pathof(Trixi) |> dirname |> dirname, "examples", "p4est_2d_dgsem")
+
+@testset "P4estMesh MPI" begin
+
+# Run basic tests
+@testset "Examples 2D" begin
+  # Note: Tests only run locally due to lack of MPI support in P4est_jll
+
+  # Linear scalar advection
+  #@trixi_testset "elixir_advection_basic.jl" begin
+  #  @test_trixi_include(joinpath(EXAMPLES_DIR, "elixir_advection_basic.jl"),
+  #    # Expected errors are exactly the same as with TreeMesh!
+  #    l2   = [8.311947673061856e-6],
+  #    linf = [6.627000273229378e-5])
+  #end
+
+  #@trixi_testset "elixir_advection_unstructured_flag.jl" begin
+  #  @test_trixi_include(joinpath(EXAMPLES_DIR, "elixir_advection_unstructured_flag.jl"),
+  #    l2   = [0.0005379687442422346],
+  #    linf = [0.007438525029884735])
+  #end
+
+  #@trixi_testset "elixir_advection_restart.jl" begin
+  #  @test_trixi_include(joinpath(EXAMPLES_DIR, "elixir_advection_restart.jl"),
+  #    l2   = [4.507575525876275e-6],
+  #    linf = [6.21489667023134e-5])
+  #end
+end
+
+end # P4estMesh MPI
+
+end # module
